Move travel certificate feature flag check to the data layer
GetTravelCertificateSpecificationsUseCase can do the check itself,
instead of always doing the network request even if the feature flag is
off, and then giving that information to the HomeItemsBuilder which must
also then check with the feature manager whether to show it or not.

Also fix providing the GetTravelCertificateSpecificationsUseCase twice
in the koin graph.