Move profile VM logic inside of it
Avoid using the mapper and do the work inline, along with testing the
ViewModel directly. This allows us to make the profile screen more
robust to specific parts of the network requests failing, while still
not showing the error screen if one thing goes wrong.
Should now just show the items that it has loaded, and allow the pull to
refresh feature to be used to retry loading the failed parts.