Add destination + handle logic to send EB value
Surface errors in the TextField itself for now.
Only validation we do for now is have the input be a non-empty string.
Fetch data only from the network on the Profile screen to always ask for
the latest EuroBonus value, and make sure to call it each time the
Profile screen is shown. Can be in the future done through optimistic
updates possible, but this is good enough for now.