Split triaging to new and legacy modules out of odyssey
Through this process, remove all the activities that those features used
 to live inside, and make them be contributed to the HedvigNavGraph in
:app.

New Claims triaging is almost there, just needs some final touches from
this point.

Bring triaging EntryPoint and ClaimGroup types along with the mutations
to a data module so that both the new and the legacy triaging modules
can access them, along with the navigation types, so that we can get
cross-module typed navigation to them.