Make triaging data module only contain the models
ClaimGroup and EntryPoint are the common models used for both legacy and
new triaging feature. These can stay in the common module, also as a way
for other modules to know about so we get proper type-safe navigation to
them. In particular, the claim flow (odyssey) feature can this way know
about the EntryPointId which it needs to start the flow.

Make the entire new triaging flow only use the first entry group
mutation, since that's all it needs in order to go through the flow and
be able to start a new claim flow.

Make the ImmutableList properties serializable through the use of a
custom Serializer, since there doesn't seem to exist a way to register
it globally.