Make AudioRecorder item adapt to the new UI
Since now it'd be white on white, a shadow is used to make it stand out
from the background colors.
Also animate the circle into the square when starting to record. Done so
by merging Recording and NotRecording composables into one, and simply
animating the different values that those two shapes need.