Adjust the claim flow screen ordering
Triage now comes in the middle of the entire flow, which is now all
encompassed in the AppDestination.ClaimsFlow graph.
This allows for the back button to work throughout, while also allowing
an easy way to "leave" the flow by popping the entire backstack
including that destination.

This moves the start claim flow logic from the pledge + notification
screen to the triage screens instead. That is because the pledge +
notification screens come before those, and we need the info from the
triage screens in order to know how to start the claims flow.

To share this information between the claim and the triage screens, the
data:claim-flow module was introduced for the shared code, hence a lot
of code has moved around.