Use HedvigBigCard all around (renamed from HedvigCardButton)
Removes the need for FormRowCard.
Adds an overload which accepts any input, but also one which is the
common case with the two texts.
Should optimally propagate the min constraints to the children to make
it easy to work with, but HedvigCard is used in many places so might as
well wait until that's done properly