Create payout specific screen
Was before contained inside the checkout screen, but was making it more
complicated than it has to be.
The logic for each thing is now split in those two destinations, in the
checkout you only see the status and pick on the payout method, and in
the payout screen the actual muatation happens, with retry capabilities
and so on.

This also moves the SingleItemCheckout screen back in the claim flow
graph, so that going back is supported properly

This also adjusts a couple of the text sizes and removes some icons from
the payout screen according to the new design.