Animate the insets consumption in the global NavHost position
Animates how we consume the insets, so that when we leave a screen which
does not consume any insets, and we enter a screen which does (by
showing the bottom nav for example) then we don't want the outgoing
screen to have its contents snap to the bounds of the new insets
immediately. This animation makes this visual effect look much more
fluid.