Remove CrossSellDetailActivity
This also removes the CrossSellFaqActivity as it's only used there

New design does not include this detail screen at all. Now simply show
them in the insurances tab, and then open the website directly.

As a consequence, the notification should now just open the insurances
tab instead of going into the detail of the cross sell.