Make moleculeViewModel store the latest emitted value and store it
The `seed` variable works as the previously named initialValue did for
the first emission, but after that we keep storing the last emission to
it. This way when we come back to the backstack entry, we know what to
emit as the first emission. We do that by giving the seed value to the
presenter, and have it do whatever it needs to do with it.
In the optimal situation, the presenter can initialize itself with some
sane defaults through this seed, and still proceed to load the data as
it would've done anyway. In the scenario where the loaded data would be
the same as the last emission, since we're backed by a StateFlow, no new
emission will happen.