Make InsuranceViewModel + Presenter use the `seed` parameter
This means that no data does not necessarily mean that we failed to load
something. Since we may get initial data from the seed and still fail to
load, resulting in a state with data + not loading but it still means
that we have failure.
To fix this, explicitly save those network errors in a mutableState.