Create InsurancePresenterTest and do changes to make it possible
Make insurance contracts return a custom type

Helps with not returning a generated model which contains GQL Fragments
etc for creating test data in tests.

Remove the android dependency `Uri` inside cross sells so that the unit
tests don't need to be android tests to have them work.

Turn the UseCases into interfaces to provide fake impls in the test.