Include all gradle modules after finding them one by one
Instead of relying on the ":path:to:module" way of defining the
directory which makes all intermediate directories also be considered a
module path, we instead walk through the project's directories inside
the "app" directory and only consider modules the directories that
contain a file with the name "build.gradle.kts".

We also manually include the "design-showcase" as that one is outside of
the "/app" directory.