Remove all unnecessary namespace declarations
Keep only two of them, "hedvig.resources" and "com.hedvig.app".
This is because hedvig.resources is the identifier we use to find our R
class to fetch our strings. And com.hedvig.app since it's used in so
many places already, specifically the R (com.hedvig.app.R) so if we feel
like it we can do that change separately.