Add class to store snooze status of the notification reminder
The reminder is supposed to show when the user does not have
notifications enabled. They can snooze this reminder which will be shown
in various parts of the app, and after some predefined time passes they
will see it again.
EnableNotificationsInfoManager is the class which supports storing this
information in persistent storage.
The UI should use this information along with whether the app currently
already has the notification permission to determine if it should be
shown or not.