Add notification reminders to profile screen
Create member-reminders-ui which provides a reusable UI for the
notifications, to be reused in all the places of the app that want to
show the member reminders.
Create notification-permission which contains the logic of requesting
the notification permission in particular. This is done so that all
places which need to show the notification permission can just reuse
this, the state object and the dialog provided by
notification-permission as a drop-in solution to have to do very little
on each call site.

Also hookup the profile screen with the rest of the possible reminders
and their actions.