Fix a bunch of animated-content related animations
The screen space taken *needs* to be full screen at all times for the
transitions to play well with each other. Otherwise the loading content
which only takes a tiny part of the screen will shrink the space taken,
and as the other content comes into the view it will animate the content
size to that new space, resulting in a very odd animation.