Improve writeNewMessage to use apollo's lock to read/write
This gets rid of our own `isSubscriptionAllowedToWrite` boolean which
was really hard to reason about, while also bringing issues with
dropping new messages under normal operation right after we've sent a
new message.

The new way we're storing the message in the cache is explained inside
the ChatRepository in the writeNewMessageToApolloCache function, which
takes care of this by using `accessCache` and doing everything within
the lambda, which is properly locked with a ReentrantReadWriteLock.

This removes the ability for us to simply drop new messages and log a
warning which we were doing before if the subscription was receiving a
new message at the same time that our "mutex" was already acquired.