Make network errors retryable and exit chat on cancellation
For the RetryChannel to still continue working after an exception is
thrown, it needs to not see the exception itself, such is the way flows
handle exceptions. If an exception is in fact thrown, then that flow is
closed forever (if there's no retry operator). This change makes it so
that the inner flow catches the exception itself, and acts accordingly
(sends the error event) and lets the retry channel continue working,
ready to receive the rety event to restart the inner flow.

On the UI side, a dialog is shown in these scenarios, which is
non-dismissable, and on positive button restarts the two messaging
flows, and on cancel simply exits the chat.

This also deletes the unused `networkError` which was never sent a value
 in the first place.