Allow molecule presenter to assert previous emissions
Instead of just doing `distinctUntilChanged` and dropping any emissions
of the exact same State, we instead do our own version of it, by
awaiting on new items when calling `awaitItem` in a loop until we
actually get a new emission. This allows us to also intercept this
process and store locally the last emitted value for later use.
With that value stored, we introduce a `awaitUnchanged` function, which
uses this last emission, but awaiting on a new item and checking that it
was in fact the same exact State result as it previously was.