Remove on launch migration of language from sharedprefs
We've had this migration since 2022-09-19. By now everyone must've run
this code to migrate their specified language. And if not, what will
happen for them is that the sharedprefs will store the old choice but it
will not be used in any way, and en-SE will be automatically chosen for
them.