Make Market non-null and use non-blocking storage
Market can always default to SE instead of being null, which does
sometimes become a problem. This persists backwards compatibility with
the market being stored in SharedPreferences, as it takes the last value
from in there, and stores it inside DataStore going forward.

We maintain the ability to get back a synchronous value of what the
market currently is by exposing a StateFlow, and always keeping it hot,
just like we've been doing with AuthTokenService and the
`val authStatus: StateFlow<AuthStatus?>` it exposes. This also comes
with the benefit that we no longer do this access on the main thread as
we did with shared preferences.