Make MarketingActivity handle locale changes well
By not recreating the activity on locale changes, we get a very smooth
experience when switching options. This still works by using
`getLocale()` which does properly read the latest configuration
composition local and updates even without the activity being recreated.

SetMarketUseCase now also properly picks the right language as the
market switches. Keeps to English if English was selected before, but
uses the local languages if that was used instead.

Due to the lack of observable APIs provided by Android for the
MarketingPresenter to listen to regarding language changes, we rely on
manually re-fetching the latest value after we know that we have changed
it.