Remove AuthTokenServiceProvider
The token provider can remain being the real one at all times. This way,
if we do end up accidentally making real network requests to the backend
when we are in demo mode we know about it as they fail completely.
Then for knowing when we are to be force logged out, we also need to
take care of checking if demo mode is on or not, something that we'd
need to check for anyway.

For all the places that depend on the AuthTokenService, they can stay
exactly as they are. Places like the FCMTokenAuthEventListener etc would
be confusing to get a fake "logged in state" from the service.