Pass Provider<T> instead of DemoOrProdProvider instance
We must do this so that all classes that require a provider can be given
a simple anonymous one for testing. On top of that, the class itself
does not need to know if the provider is the DemoOrProd one or whatever
else it may be.
HomePresenterTest is an example of how this change helps just do
`Provider { T() }` in tests with this change.