Add DatadogAttributesManager to handle datadog attributes
Currently, to keep track of some user info which helps us fix bugs, we
need to add items both on Datadog and on GlobalRum objects. This class
will serve both as a way for this to be simplified and be done in one
place, while also allowing outside modules to contribute their own
attributes.