Split AuthEventStorage out of AuthEventBroadcaster
This way, the ones emitting the events can just emit to
AuthEventStorage which gets initialized without any dependencies.
Then we have the AuthEventBroadcaster simply consume those events, and
we make it initializable, and let it do its own thing for the entire
lifetime of the application.

This breaks the ever-repeating problem of AuthEventListeners needing
access to classes like ApolloClient OkHttp but in turn those need all
the listeners to already be initialized.