Fix Activity -> Fragment migration bugs
As we go to Fragment from Activity, some overrides need to be changed
and most importantly, listeners that were attached to a "bigger" scope
than what the fragment lives in need to be cleared out on destroy of the
view.

So remove the onGlobalLayout listener on destroy view and only create a
new one onCreateView so that it has the latest reference to the
`binding` object.