Force navigation for navigation events
When we use navigate with the backstackentry lifecycle in mind, we do it
so that we do not have navigation events overlapping and going to the
same navigation destination twice.
This works great when we do navigation events due to some click action.
However, this does *not* work great for situations where the navigation
event comes once already anyway due to some state change and so on. That
is because this navigation may get swallowed if a previous navigation
event was happening already.

Particularly in the cancel insurance case, what was going on is that as
we are going into the termination flow, we immediately do a network
request to know where to go next. This event comes in, but we are still
navigating from the previous screen, so this nav event *sometimes* just
gets swallowed. This change fixes this, and we must consider other
situations where navigation happens due to a state change, and not due
to a user interaction, that we use the unsafe variant of the Navigator