Use the provider for GetInsuranceContractsUseCase
Now with the Provider<> being present in some places, there is the risk
that we forget to provide the right dependency through DI to it.
In this case, the class was still expecting just T instead of
Provider<T> which was crashing the app. That is because we no longer
provide an instance of type T to the DI graph. We provide TImpl, TDemo
and the Provider<T> through the ProdOrDemoProvider<T>.

An easy mistake to do, and we are not really guarded by any type-safety
here, besides actually running the app and getting the runtime crash.