Fix moving flow back handling
Make back handling move normally through the backstack instead of
exiting the entire flow in some scenarios.
For this to work, instead of relying on some state being there for the
navigation to trigger, instead handle that through event triggers baked
in the ui state itself which the UI clears when the navigation event is
properly handled.
This allows the member to go back and forth to all moving flow
destinations without having to exit the entire flow for no reason.
Going forward again triggers the same path as before, overriding the old
results in case of a new success response from the backend.
Remove all BackHandler calls to support this.
Keep ChangeAddressDestination.AddressResult outside of the
AppDestination.ChangeAddress navigation graph, and make sure that when
we do end up there, we do it after popping the entire backstack up to
ChangeAddressDestination.AddressResult.