Tiny UI fixes for offer
Allow HedvigInfoCard take in a new shape if the caller needs it to
Make the small info card use the right squircle corner radius
For cases where the divider is added on a list on all but the last item,
prefer adding it in the first line by doing a much simpler `index > 0`
check instead.
Remove some non lokalised content description.
Use the right icon for the info icon
Use the right tint for the arrow showing that there is more info in the
quote details card
Remove the odd color for the ExpandablePlusCard which made it look
different from the cards above, especially in dark mode. New colors
match Figma instead.