Fix overly verbose error logs
The real solution here is for `safeExecute` to report properly errors
that are expected (like being unauthenticated) versus ones that are real
problems and need to be logged with Error priority.
This work is postponed till the safeExecute function is able to just
return the correct Either directly, when no more callers rely on it
being OperationResponse. That way we will then also remove all
`toEither` calls at the same time.

On top of this, the logging for the error cases can happen globally in
there too, to get better logs, possibly along with the query name,
parameters and so on without having to opt-in for each caller.
Extra logs can then still happen at call-site when there's such a need,
but not to get the basic error reporting.