Fix home screen jumpiness
Simply do not render the layout until we get the screen size. This way
the HomeLayout won't first assume the screen size has 0 height and try
to lay out everything in a column at the top of the screen.
This way the animated placement on the cards won't do the jump that it
used to