Finish connect-payment implementation
Bump authlib and use the exposed PaymentRepository to fetch the right
payment connection URL from there directly
Add adyen screen and Presenter
Add a special error case for when the initialization of connect payment
fails, and not the connecting of the payment itself.
Fix the loading bar in the WebViews by making it go on top of the
WebView instead of pushing it when it appears.