Add basis for new chat UI + Functionality
Can at this stage:
send messages
poll for latest messages
ask for more messages when reaching the top
Preview the message which is still being sent
Collapse the top app bar when scrolling upwards and show it scolling
downwards
Scroll to the bottom when sending a new message or opening the keyboard
Scroll to the bottom when receiving a new message and is already
looking at the last message before
Click on messages that contain links and open them. This also handles
deep links that are directed into our own app.