Adjust size of image in the chat according to its width/height ratio
Adds a special scenario for when a message only contains a link to a
gif and tries to render it as a gif instead of a message.
Adjusts the width/height of it according to its intrinsic width/height
ratio and makes it take an adequate size in the chat space in order to
be easy to read.