Fix photos not sending when coming back to a killed StateFlow
Now when the new stateFlow becomes hot, it can continue consuming the
events from inside the Channel which stay there until a consumer pops
up, instead of dropping the event when there are no consumers.

Ensure that the chat messages are still there when coming back into this
restored state situation by reading the old ones from `lastState`.

The situation where the entire process is killed in the background while
still taking a picture is handled by `rememberChatFileState` which uses
the ChatFileState.Saver to properly restore the state in that scenario,
and still remember the URI to the picture which was just taken.