Fix Retrofit service crashing on startup
By default the interface will be used to auto-generate code from
Retrofit for each one to an endpoint. Here the `uploadFile` is just a
convenience to call the real function, so extracting it to an extension
function means that retrofit won't try to auto-generate code for it
which shouldn't be there in the first place.