Add a button to send an existing photo/video in the chat
To fetch the photo/video, we use the built-in photo picker, which is
also backported to older API versions, making the experience as seamless
as it can get.

This differentiates sending a photo taken on the fly, to sending some
media picked from the existing phone storage. These two are handled
differently since for photos taken on the fly we can get a File handle
to them in order to upload them. But for the URI received from the photo
picker, we just read its contents through the input stream using
contentResolver instead. That is done through a custom RequestBody
passed to Retrofit.

Also add a mime type for mp4 to consider rendering them in a special
manner in the future.