Stop displaying messages that are still being sent
Since we do not have a way to be able to match the message we will be
sending with a message that we will later be receiving, because the ID
is defined from the backend, we instead stop trying to handle this
scenario and instead do the following.
When any type of message is sent, not only text messages as before, we
take the immediate response back and store it into the cache
immediately, not waiting for the polling to receive it. This should
provide quick feedback after the message is sent.
To handle failed scenarios, we do in fact keep track of those locally in
that case, and we can use that fact to allow members to retry sending
them.