Allow the member to retry sending failed messages
Currently only works with an in-memory list of failed attempts, which
can only be retried at that moment, and are lost if the chat is closed.
Also does not render local URIs with their actual content, even if they
are pictures, but instead shows a "file attached" message, with the
ability to retry by clicking on it, in the same way that the text
message can be retried.