Make failed to be sent Uri to render in a special way
If the Uri can just be rendered as is, Coil can take care of rendering
by taking the Uri from the device directly. If it fails to render it,
due to it being a Uri pointing at a pdf file for example, then a
fallback painter is introduced which indicates that it's possible to be
retried. When clicked, in either case, we try to send that content
again.