Continuously automatically try to fetch more when at the top of the chat
Instead of having an ugly retry button at the top, we can simply always
keep trying to fetch more as long as we are scrolled all the way to the
top. If they member scrolls down again, the composable gets disposed and
we simply do not try to fetch more until it comes in composition again.