Fix initial emission silently failing when cache is clear
This was problematic in scenarios where one enters the chat while the
cache is empty, and there is no internet connection, which resulted in
the cache never being populated, and therefore the chat always
considering itself still initializing forever until internet is back.
Now an empty list is successfully emitted, and when internet is back the
messages will still automatically come, but now we allow the member to
interact with the chat as normal.