Get rid of all unused theme/xml values in XML
Simplify themes into one for splash and one for the normal theme. And
the colors are simply a black and a white, and they are used for the
splash screen and for the launched icon.

Also get rid of some custom v27/v29 theming, as that will be handled
from `enableEdgeToEdge()` instead. Now we only need a tiny bit of
configuration around dark/light theme since the splash screen needs to
support that and can not read the user preference at that point.