Update LoggedInActivity to work with darkTheme from HedvigAppState
When splash screen is showing, it's too early to be able to read the
user-selected dark/light mode selection, so that shows the color
following the system. But that then interferes with `enableEdgeToEdge`
since it will try to add a scrim in order to make the icons legible then
when we make the background be of the opposite color. This happens in
situations where we are in light mode, so splash screen shows as white
background, but the app has a forced dark mode internally.

To combat this, when the splash screen is removed we signal that to
re-run `enableEdgeToEdge()` after the splash is gone, so that it sees
the new background, and it does not put that extra scrim color for no
reason, and the status bar colors are updated properly.

We still call AppCompatDelegate.setDefaultNightMode just to be safe,
even though we may not actually be using this in some way in the app.