Cleanup setting language and market
Before, sometimes on startup there'd be no language properly set and
we'd default back to en_SE, despite someone already having chosen a
different language, but in a way that was not understood by the
"migration" function.

Now we always run a "migration" function on startup, checking the
current market and setting the right language directly, or defaulting to
SE market if none was selected, and consequently selecting a language
in the process.

This should now work for all API levels in a consistent way, which will
not be cleared after the app is closed and started again.