Fix chat messages scrolling to the top on initialization
Initially, when there is nothing in the lazy list, the messages are
empty, and what shows in the lazy list is the loading indicator and a
spacer here.
https://github.com/HedvigInsurance/android/blob/c1d80d089e9cf7af184e94503fddcc09c70f67f5/app/feature/feature-chat/src/main/kotlin/com/hedvig/android/feature/chat/ui/ChatLoadedScreen.kt#L277-L308

Then after the messages show, what LazyList tries to do when new items
come into the list is to see the unique ID of each of the items that
were showing before, and still keep those items in view, because
otherwise it looks super odd to have new items come in and suddenly what
you were looking at just disappears.

Now what that means is that all the chat messages come first in the
list, but it tries to keep the loading indicator in the screen too, so
despite the fact that we have reverseScrolling set to true and it should
scroll from bottom-up, it just scrolls all the way to the top in order
to keep those two items at the top.

Not adding any items to the list before the first messages show up fixes
this issue.

To still show a loading state, we show the same loading indication as we
did when the chat is still initializing until the first messages come in
since we can not rely on the loading indicator to show the loading state