Make expanded content always take up the entire width
The Markdown composable as it seems for the first frame takes up no
space. This means that since we place it inside an AnimatedVisibility
composable, the moment it gets the available width it animates the width
too, from being 0.dp to the width that it needs to be.
This change makes it always take the entire width, so the text comes in
immediately without animating on the X axis.