Fix termination flow navigation issues
Remove all BackHandlers and custom finishing of the destinations
Instead hookup the graph properly, with a start destination starting the
flow with user input on the button, and going from there. When reaching
terminal destinations like the success, failure or the unknown screen,
then we still make sure that as we navigate we remove the nested nav
graph from the backstack. That way going back always does the right
thing, same with navigating up.