Store successful member requests to delete their account
Can be used as a local way to be aware if someone has already requested
an account deletion. This is stored only after we know our call to the
backend was done successfully, and is done so since we do not keep track
of the accounts pending deletions on the backend.
We keep track of this by storing their member-id so that one can
avoid accidentally getting this status when logging in from another
account on the same device.