Make FilesLazyVerticalGrid a standalone composable and fix the UI
Removes the extra work of adding the buttons at the bottom along with it
being an extension on ColumnScope. This allows it to be used more
easily from any kind of layout, while respecting the `modifier` passed
into it to be applied on the top level composable instead of on a subset
of the composables called inside of it.

Change the X button to go outside of the card bounds by using
wrapContentSize, which results in the spacing between the items to not
be ruined because of it existing. Adjust the ContentPadding passed into
the lazy grid by default to adjust for this change so that the button
and the shadow do not get clipped.

Adjust the lazy grid to try to fit the size from figma by adding more
than 3 items on the grid if the size allows for it, while still keeping
the 1:1 aspect ratio