Adapt to the new material3 APIs
Namely the date picker now asks for a locale, which we get from the
LanguageService.
The validator used to allow or disallow specific dates is now specified
on the date picker state directly through the SelectableDates interface
instead, which does simplify us not having to pass the validator all the
way down to the UI which calls the DatePicker.