Fix top app bar using insets from the bottom
WindowInsets.systemBars does properly take into account to only use the
horizontal and the top insets. The change to use
WindowInsets.systemBars here did not account for it, so this needs to
be reverted back.
TopAppBarDefaults.windowInsets is using
WindowInsets.systemBars.only(WindowInsetsSides.Horizontal + WindowInsetsSides.Top)
anyway, and we add the cutout to it as mentioned here
https://issuetracker.google.com/issues/326526714