Fix the screen clipping its size when navigating away from UrlFetched
When we were in the `UrlFetched` state of this screen, there was no UI
emitted. This actually ends up making the content of the NavHost go down
to showing no UI, so since the entire space is not needed, it only takes
up 0 pixel and clips the rest of the content. This is done with a center
alignment by default in the NavHost. This results in the when navigating
away from this screen, the next screen as it animates into view, it also
animates the content size accidentally, expanding from the center
towards the edges, making it look like it's clipped.
This wraps the entire screen in a simple box which fills max size to
ensure that no matter which state the screen is in, it always takes up
the entire screen size as it should, and this fixes the navigation
clipping issue.