Tweak the label scaling more to handle edge cases
Pass in just a string for the label, so that inside the component we can
force the text to be max 1 line and clip on overflow by default.
Also force the font size density to be capped to a reasonably value that
was eyeballed to work on very small screens with the font size scaled to
the maximum.
We also still allow the text to go a bit over to the item next to it,
with the idea being that only the insurance text should be doing that in
all scenarios, so this should for over 99.9% of our users not overlap at
all.
Not a perfect solution by any means, but no matter what, a trade-off
needs to happen here.