Make Payments screen continue rendering the existing data on back
Fix clearing the data on refresh when the presenter starts presenting
Remove unnecessary events and types that were only used in other screens
We still keep the `contentForOtherScreens` because of the way these
screens are structured for now. But in the future each screen should
fetch their own information instead.
Map the raw PaymentOverviewData into a screen-specific model to make the
code in the UI much simpler to go through rather than checking for all
the deep parameters inside a big object making the UI code harder to
read.