Fallback to system back if popBackStack does nothing
Enhance the behavior inside the `Navigator` which all screens are using
to pop the backstack when a button which does not do `navigateUp()` but
still wants to go `back` one screen is pressed.
This fixes the problem as described here
https://issuetracker.google.com/issues/329046732 which makes it so that
if you end up in a destination without anything else on your backstack,
which typically happens when you are in the app from a deep link. Now
instead of that button doing nothing, and even breaking future
`navigateUp()` calls, it will fall back to whatever behavior the
system's back button has, which in the deep link scenarios it is to go
back to the app which they came from.