Update failed popBackStack calls to finish the app directly
The documentation actually explicitly calls out this use case and asks
to use finish instead of trying to dispatch a back press to the
onBackPressedDispatcher. This is so that if there was some in-between
dispatcher attached for whatever reason, we'd actually not be getting
the behavior we want.
More context here:
https://developer.android.com/guide/navigation/backstack#handle-failure