Stop allowing to submit the wrong number of ssn digits
Otherwise since both DK and NO use the same login method one could try
to login to DK when they've picked a NO locale (but not the other way
around due to the max digit limit, by coincidence).
This disallows sending it in the first place unless the ssn size is just
right.