Delete the MarketingActivity completely
Simply merge its graph into the existing NavHost to make them part of
the same hierarchy and be able to navigate from one another.

Add functions inside HedvigAppState to be able to go from and to the
loggedin/loggedout scenarios. For now without saving/restoring state
due to https://issuetracker.google.com/issues/334413738

Simplify ActivityNavigator which does not need to do this work anymore

Move the logic of observing the auth status to go to the logged out
activity to be done as a compose effect and navigate to the LoggedOut
nav graph instead. This allows us to observe the demo mode and
AuthStatus when the app is at least Started and combine that with the
current app's destination to know if we need to take action or not.

Also move functions like the openEmail just into this one activity which
existed inside the marketing activity before