Delete ForceUpgradeActivity and move logic to LoggedInActivity
This gets rid of the last non-main Activity in the Hedvig app!

The entire app is wrapped in an if-else on the kill-switch if that ever
turns to true. This is done so that the entire NavHost is removed from
the backstack and we still handle the back gesture properly which will
at that point take the member out of the app.

Move the logic of fetching and caching the kill-switch flag to the
HedvigAppState