Re-create the button component from design
Exposes a hopefully simple API to chose between the four sizes and the
five styles on each call site. Sets the right defaults automatically by
fetching the right information from our tokens to do so.
Duplicate some code around the ThreeDotsLoading switch in the loading
state, but this should eventually be the only place which has its use
anyway.