First almost working iteration of the text field
Fix the API to take a sealed hierarchy for all possible error scenarios
Reusing a lot of the m3 text field layout logic here, despite probably
not necessarily needing all of it. But this is giving us the label
that we do need here.
Future steps would be to strip away even more unnecessary parts here