Make all of the text field part of the transition scope
This now allows for the entire layout to adjust itself in all states.
This also helps with the padding values being different in label or no
label scenarios so that if the size would change but in the two end
positions there would be no change, the container layout does in fact
stay put at its place without moving unnecessarily.