Use custom layout for textfield
We need the behavior of weight(1f) on the textfield so that the leading
and trailing contents always show, yet we do not want the textfield to
take up 0.dp width if we are in a horizontally scrollable layout
since that's what weight will do there.
RowTextFieldLayout is a specialized mix of the two for our specific use
case.